refactor: 관심사 분리 - 비즈니스/상태변환 & UI 부분으로 분리
